wic64.asm: Slightly rearranged code to fix optimisation issues
Recently added code caused time critical sections to move in a way that
would cause a branch to cross a page boundary, adding another cycle per
transferred byte and thus decreasing transfer speed by about 3kb/s.

Rearranged code to perform like before, that is when the library
is included at or close to a page boundary, e.g. $0810.

Might add an explicit optimization option that aligns critical sections
so that the best possible performance can be guaranteed regardless of
the address at which the library is included.