wic64.asm: Fixed timeout handler failure when changing direction
The code to change the transfer direction can not be implemented in
a separate subroutine that is called from wic64_receive_header. All
invocations of the timeout handling code must be on the same stack
level from the calling code for the stack manipulation that allows a
virtual branch to the user timeout handler to work correctly.

This caused the wic64-update program to be temporarily broken, since the
expected timeout of the update command caused by rebooting the esp before
sending a response occurs when trying to change transfer direction on the
client.