tools/export.rb: Added export to dasm format, restructured code
The export script uses da65 to dissassemble the assembled binary and to
transform the code to dasm format, exporting the symbols necessary to
use the implementation from dasm-based code.

In wic64.asm, labels that should be exported are marked with a magic
comment on the same line containing the tag "EXPORT".

The makefile target "wic64.dasm" transpiles the code to dasm in a
position independent way suitable for inclusion, while the target
"verify-dasm-export" assembles the transpiled version as a standalone
binary using dasm, and finally diffs it against the acme binary to make
sure both are identical.

The dasm-Version is intended to be used to develop a wic64 library for
XCBasic, which is a BASIC dialect that transpiles and assembles using dasm.

In order for da65 to properly dissassemble the data section containing
the runtime variables into .byte directives instead of nonsensical code,
all runtime variables have been moved into a consequtive section.

Also added some diagnostic, marking the performance critical sections
and issuing a warning during assembly if those sections cross page
boundaries.