wic64.asm: Make sure default nmi is set up during wic64_load_and_run
Since the .receive_and_run part of wic64_load_and_run does not include
timeout detection, make sure the user can at least press runstop/restore
in case the code gets stuck in an infinite loop due to a timeout
condition. To achieve this, the kernal is banked in and the NMI vector
is reset to its default value at the beginning of .receive_and_run.

Added/improved comments as well.