wic64.asm: Fixed bug in wic64_send_header breaking timeout handler
Can't use jsr wic64_send to send the header since then the stack level
is not consistent with the other toplevel api routines when a timeout
occurs while sending the header (one level deeper). Send the header
in a short loop instead.

This also makes the .dont_update_transfer_size_next_time obsolete, which
was some kind of an ugly hack anyways.