wic64.asm: Improved wic64_detect and fixed it for dasm-export
To detect the legacy firmware, also test for 14 bytes in addtion to
13 bytes. The fimrware string response is 13 bytes, but earlier versions
might respond with "Command error." (14 bytes) as well.

In order to avoid having to reserve 30 bytes to receive the response,
the response is now manually received and discarded.

Also don't use wic64_branch_on_timeout to avoid creating position
dependent code that would break the export to dasm format.