wic64.asm: Simplified wic64_detect
Since the legacy firmware always receives the whole request, we can
determine that a wic64 is present when send_header does not time out.
Now sends a standard protocol version request and receives the response
without caring for timeouts anymore. If the response size is within
the expected range, a current firmware has responded, if the response
size doesn't change, a legacy firmware has not responded since it did
not understand the request.

This is simpler than jumping through hoops to craft a request that both
firmwares still understand.