wic64.h: Added macro +wic64_execute <request> (response is ignored)
This variant of wic64_execute discards the response payload data by
temporarily modifiying the store instruction to a nop.

Useful for commands that never send a response payload, or in cases where
the response payload can be ignored. Avoids having to reserve memory for
a response that is never used.