wic64.h: Don't discard response if store instruction is explicitly set
When wic64_execute is called without a response address, only discard the
response if the store instruction has not already been changed, i.e don't
overwrite the users custom instruction.

This allows the user to omit the response address either because the
response is irrelevant, or because the user knows that the response will
be handled by the custom store instruction. Specifying a response address
is pointless in either case.

Also use constants for opcodes instead of literal values to improve
readability.