wic64.asm: Added automatic discard of response, fixed single arg execute
Fixed automatic discard of response data for +wic64_execute <request>
variant. Introduced wic64_auto_discard_response to control whether
the store instruction should be set/reset to a nop instruction from
initialize/finalize instead of doing this directly in the macro.

This is necessary because otherwise the store instruction might not get
reset if custom handlers where installed and timeouts or errors occurred.
Handling this in initialize makes sure the automatic discard is always
removed again in finalize, and also keeps the macro code simple.