wic64.h: Fixed size argument handling in partial send/receive macros
Don't add 1 to the value of the highbyte of the size .argument. This is
a stupid mistake on my part, which should have been caught by the test
case for transfers in discrete steps, but it seems there is a bug in acme
that masked this specific bug. The bug in these macros becomes apparent
when using C64Ass instead. Need to investigate further.