wic64.h: Added variants of send/receive macros with indirect size argument
Using acme's pass-by-reference feature for macro arguments, the size
argument for the wic64_send/receive macros can now alternatively be
passed indirectly, i.e. as the address in memory where the actual size
value is stored, by prefixing it with a tilde character

For example, given

size: !word $1040
destination:

then

+wic64_receive destination, ~size

is equivalent to

+wic64_receive destination, $1040

with the difference that the actual size can be changed at runtime
instead of hard-coding it at assembly time.